+++ src/misc.rs
@@ -1,0 +1,19 @@
+//! Miscellaneous pthread functions
+
+pub fn init() {}
+
+#[no_mangle]
+pub unsafe extern "C" fn sched_yield() -> libc::c_int {
+    ctru_sys::svcSleepThread(0);
+
+    0
+}
+
+#[no_mangle]
+pub unsafe extern "C" fn pthread_sigmask(
+    _how: ::libc::c_int,
+    _set: *const libc::sigset_t,
+    _oldset: *mut libc::sigset_t,
+) -> ::libc::c_int {
+    -1
+}
